1. remove useless field checks.status and modify according scripts 2. clean-up BasicFunctions.class.php to have all functions listed in order 2. fix bugs: (1) javascript document.form.submit() not work on 'edit check' and 'edit guideline' pages (2) delete a guideline without any groups/subgroups hit error: rows is not an array (3) remove left-over test data from language_text.sql

+++ include/classes/DAO/GuidelineGroupsDAO.class.php
@@ -155,8 +155,11 @@ class GuidelineGroupsDAO extends DAO {
 		         WHERE group_id = ".$groupID;
 		$rows = $this->execute($sql);
 		
-		foreach ($rows as $row)
-			$guidelineSubgroupsDAO->Delete($row['subgroup_id']);
+		if (is_array($rows))
+		{
+			foreach ($rows as $row)
+				$guidelineSubgroupsDAO->Delete($row['subgroup_id']);
+		}
 		
 		// delete language for group name
 		$sql = "DELETE FROM ".TABLE_PREFIX."language_text
